Solvers

Solvers

Minimizing a problem

StructuredOptimization.@minimize — Macro.

@minimize cost [st ctr] [with slv_opt]

Minimize a given problem with cost function cost, constraints ctr and solver options slv_opt.

Example

julia> A, b = randn(10,4), randn(10);

julia> @minimize ls(A*x-b) + 0.5*norm(x);
    it |      gamma |        fpr |        tau |        FBE |
 ------|------------|------------|------------|------------|
     1 | 2.9152e-02 | 2.7656e+00 | 1.0000e+00 | 5.5181e+00 |
     9 | 2.9152e-02 | 9.9682e-05 | 1.0000e+00 | 4.4086e+00 |

julia> @minimize ls(A*x-b) st x >= 0.;
    it |      gamma |        fpr |        tau |        FBE |
 ------|------------|------------|------------|------------|
     1 | 5.8304e-02 | 1.0068e+00 | 1.0000e+00 | 6.6282e+00 |
     3 | 5.8304e-02 | 9.5210e-16 | 1.0000e+00 | 6.5654e+00 |

julia> it, slv = @minimize ls(A*x-b) st norm(x) == 2.0 with PG(maxit = 5);
    it |      gamma |        fpr |
 ------|------------|------------|
     1 | 6.1373e-02 | 2.2090e+01 |
     5 | 3.0686e-02 | 5.5190e-01 |

Returns as output a tuple containing the number of iterations and the constructed solver.

source

Problem warm-starting

By default warm-starting is always enabled. For example, if two problems that utilize the same variables are solved consecutively, the second one will be automatically warm-started by the solution of the first one. That is because the variables are always linked to their respective data vectors. If one wants to avoid this, the optimization variables needs to be manually re-initialized before solving the second problem e.g. to a vector of zeros: ~x .= 0.0.

Specifying solver and options

As shown above it is possible to choose the type of algorithm and specify its options by creating a Solver object. Currently, the following algorithms are supported:

Proximal Gradient (PG) [1], [2]

Fast Proximal Gradient (FPG) [1], [2]

ZeroFPR [3]

PANOC [4]

StructuredOptimization.PG — Type.

PG(;kwargs...)

Creates an object PG containing the options of the Proximal Gradient solvers:

gamma, stepsize (default: unspecified, determined automatically)

maxit, maximum number of iteration (default: 10000)

tol, halting tolerance on the fixed-point residual (default: 1e-4)

adaptive, adaptively adjust gamma (default: false if gamma is provided)

fast, enables accelerated method (default: false)

verbose, verbosity level (default: 1)

verbose_freq, verbosity frequency for verbose = 1 (default: 100)

source

StructuredOptimization.FPG — Function.

FPG(;kwargs...)

Same as PG, creates the options of the Fast Proximal Gradient solver.

source

StructuredOptimization.ZeroFPR — Type.

ZeroFPR(;kwargs...)

Creates an object ZeroFPR containing the options of the ZeroFPR solver:

gamma, stepsize (default: unspecified, determined automatically)

maxit, maximum number of iteration (default: 10000)

tol, halting tolerance on the fixed-point residual (default: 1e-4)

adaptive, adaptively adjust gamma (default: false if gamma is provided)

fast, enables accelerated method (default: false)

verbose, verbosity level (default: 1)

verbose_freq, verbosity frequency for verbose = 1 (default: 100)

memory, memory of the LBFGS operator (default: 10 )

source

StructuredOptimization.PANOC — Type.

ZeroFPR(;kwargs...)

Creates an object PANOC containing the options of the PANOC solver:

gamma, stepsize (default: unspecified, determined automatically)

maxit, maximum number of iteration (default: 10000)

tol, halting tolerance on the fixed-point residual (default: 1e-4)

adaptive, adaptively adjust gamma (default: false if gamma is provided)

fast, enables accelerated method (default: false)

verbose, verbosity level (default: 1)

verbose_freq, verbosity frequency for verbose = 1 (default: 100)

memory, memory of the LBFGS operator (default: 10 )

source

Build and solve

The macro @minimize automatically parse and solve the problem. An alternative syntax is given by the function problem and solve.

StructuredOptimization.problem — Function.

problems(terms...)

Constructs a problem.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> p = problem(ls(A*x-b), norm(x) <= 1)

source

StructuredOptimization.solve — Function.

solve(terms::Tuple, solver_opt::ForwardBackwardSolver)

Takes as input a tuple containing the terms defining the problem and the solver options.

Solves the problem returning a tuple containing the iterations taken and the build solver.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> solve(p,PG());
    it |      gamma |        fpr |
 ------|------------|------------|
     1 | 7.6375e-02 | 1.8690e+00 |
    12 | 7.6375e-02 | 9.7599e-05 |

source

It is important to stress out that the Solver objects created using the functions above (PG, FPG, etc.) specify only the type of algorithm to be used together with its options. The actual solver (namely the one of ProximalAlgorithms.jl) is constructed altogether with the problem formulation. The problem parsing procedure can be separated from the solver application using the functions build and solve!.

StructuredOptimization.build — Function.

build(terms::Tuple, solver_opt::ForwardBackwardSolver)

Takes as input a tuple containing the terms defining the problem and the solver options.

Returns a tuple containing the optimization variables and the built solver.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> p = problem( ls(A*x - b ) , norm(x) <= 1 );

julia> build(p, PG());

source

StructuredOptimization.solve! — Function.

solve!( x_solver )

Takes as input a tuple containing the optimization variables and the built solver.

Solves the problem returning a tuple containing the iterations taken and the build solver.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> p = problem( ls(A*x - b ) , norm(x) <= 1 );

julia> x_solver = build(p, PG(verbose = 0));

julia> solve!(x_solver);

source

References

[1] Tseng, On Accelerated Proximal Gradient Methods for Convex-Concave Optimization (2008).

[2] Beck, Teboulle, A Fast Iterative Shrinkage-Thresholding Algorithm for Linear Inverse Problems, SIAM Journal on Imaging Sciences, vol. 2, no. 1, pp. 183-202 (2009).

[3] Themelis, Stella, Patrinos, Forward-backward envelope for the sum of two nonconvex functions: Further properties and nonmonotone line-search algorithms, arXiv:1606.06256 (2016).

[4] Stella, Themelis, Sopasakis, Patrinos, A simple and efficient algorithm for nonlinear model predictive control, 56th IEEE Conference on Decision and Control (2017).