Solvers

Solvers

Minimizing a problem

StructuredOptimization.@minimize — Macro.

@minimize cost [st ctr] [with slv_opt]

Minimize a given problem with cost function cost, constraints ctr and solver options slv_opt.

Example

julia> A, b = randn(10,4), randn(10);

julia> @minimize ls(A*x-b) + 0.5*norm(x);
    it |      gamma |        fpr |        tau |        FBE |
 ------|------------|------------|------------|------------|
     1 | 2.9152e-02 | 2.7656e+00 | 1.0000e+00 | 5.5181e+00 |
     9 | 2.9152e-02 | 9.9682e-05 | 1.0000e+00 | 4.4086e+00 |

julia> @minimize ls(A*x-b) st x >= 0.;
    it |      gamma |        fpr |        tau |        FBE |
 ------|------------|------------|------------|------------|
     1 | 5.8304e-02 | 1.0068e+00 | 1.0000e+00 | 6.6282e+00 |
     3 | 5.8304e-02 | 9.5210e-16 | 1.0000e+00 | 6.5654e+00 |

julia> it, slv = @minimize ls(A*x-b) st norm(x) == 2.0 with PG(maxit = 5);
    it |      gamma |        fpr |
 ------|------------|------------|
     1 | 6.1373e-02 | 2.2090e+01 |
     5 | 3.0686e-02 | 5.5190e-01 |

Returns as output a tuple containing the number of iterations and the constructed solver.

source

Problem warm-starting

By default warm-starting is always enabled. For example, if two problems that involve the same variables are solved consecutively, the second one will be automatically warm-started by the solution of the first one. That is because the variables are always linked to their respective data vectors. If one wants to avoid this, the optimization variables needs to be manually re-initialized before solving the second problem e.g. to a vector of zeros: ~x .= 0.0.

Specifying solver and options

You can pick the algorithm to use as Solver object from the ProximalAlgorithms.jl) package. Currently, the following algorithms are supported:

ProximalAlgorithms.ForwardBackward, also known as proximal gradient

method [1], [2]. Nesterov acceleration can be enabled, which significantly improves its performance for convex problems.

ProximalAlgorithms.ZeroFPR, a Newton-type forward-backward algorithm,

proposed in [3], using L-BFGS directions to accelerate convergence.

ProximalAlgorithms.PANOC, another Newton-type forward-backward algorithm,

proposed in [4], also using L-BFGS directions.

Parse and solve

The macro @minimize automatically parse and solve the problem. An alternative syntax is given by the function problem and solve.

StructuredOptimization.problem — Function.

problems(terms...)

Constructs a problem.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> p = problem(ls(A*x-b), norm(x) <= 1)

source

StructuredOptimization.solve — Function.

solve(terms::Tuple, solver::ForwardBackwardSolver)

Takes as input a tuple containing the terms defining the problem and the solver options.

Solves the problem returning a tuple containing the iterations taken and the build solver.

Example

julia> x = Variable(4)
Variable(Float64, (4,))

julia> A, b = randn(10,4), randn(10);

julia> p = problem(ls(A*x - b ), norm(x) <= 1);

julia> solve(p, ProximalAlgorithms.ForwardBackward());

julia> ~x
4-element Array{Float64,1}:
 -0.6427139974173074
 -0.29043653211431103
 -0.6090539651510192
  0.36279278640995494

source

Once again, the Solver objects is to be picked from ProximalAlgorithms.jl).

References

[1] Tseng, On Accelerated Proximal Gradient Methods for Convex-Concave Optimization (2008).

[2] Beck, Teboulle, A Fast Iterative Shrinkage-Thresholding Algorithm for Linear Inverse Problems, SIAM Journal on Imaging Sciences, vol. 2, no. 1, pp. 183-202 (2009).

[3] Themelis, Stella, Patrinos, Forward-backward envelope for the sum of two nonconvex functions: Further properties and nonmonotone line-search algorithms, arXiv:1606.06256 (2016).

[4] Stella, Themelis, Sopasakis, Patrinos, A simple and efficient algorithm for nonlinear model predictive control, 56th IEEE Conference on Decision and Control (2017).